pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "thakaamed-dicom"
version = "1.0.0"
description = "THAKAAMED DICOM Anonymizer - Professional medical imaging de-identification for Saudi healthcare"
readme = "README.md"
license = {text = "CC BY-NC-ND 4.0 (Non-Commercial Research Only)"}
requires-python = ">=3.10"
authors = [
    { name = "THAKAAMED AI", email = "info@thakaamed.com" }
]
maintainers = [
    { name = "THAKAAMED AI", email = "dev@thakaamed.com" }
]
keywords = [
    "dicom",
    "anonymization",
    "de-identification",
    "medical-imaging",
    "healthcare",
    "hipaa",
    "gdpr",
    "pdpl",
    "saudi-arabia",
    "vision-2030",
    "thakaamed"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Private :: Do Not Upload",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "pydicom>=2.4",
    "reportlab>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "types-PyYAML",
]

[project.scripts]
thakaamed-dicom = "thakaamed_dicom.cli.main:main"

[project.urls]
Homepage = "https://github.com/thakaamed/dicom-anonymizer"
Documentation = "https://thakaamed.github.io/dicom-anonymizer"
Repository = "https://github.com/thakaamed/dicom-anonymizer"

[tool.hatch.build.targets.wheel]
packages = ["src/thakaamed_dicom"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (needed for Click)
]

[tool.ruff.lint.isort]
known-first-party = ["thakaamed_dicom"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/thakaamed_dicom"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
